requirements.txt
asgiref==3.8.1
boto3==1.38.12
botocore==1.38.12
crispy-bootstrap4==2024.10
crispy-bootstrap5==2025.4
Django==5.1.7
django-crispy-forms==2.3
django-environ==0.12.0
django-storages==1.14.6
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
